Remove desktop folder.
