+++ .testcaferc.json
@@ -1,5 +1,5 @@
 {
-  "src": "test/functional/specs/desktop/C2581.js",
+  "src": "test/functional/specs/*.js",
   "skipJsErrors": "true",
   "clientScripts": [
     { "path": "test/functional/helpers/alloyPageSnippet/index.js" },
