+++ test/functional/specs/C2560.js
@@ -1,6 +1,6 @@
 import { t, ClientFunction } from "testcafe";
-import fixtureFactory from "../../helpers/fixtureFactory";
-import testServerUrl from "../../helpers/constants/testServerUrl";
+import fixtureFactory from "../helpers/fixtureFactory";
+import testServerUrl from "../helpers/constants/testServerUrl";
 
 const getAlloyFunction = ClientFunction(() => !!window.alloy);
