+++ test/functional/specs/C2595.js
@@ -1,22 +1,22 @@
 import { t, ClientFunction } from "testcafe";
-import createNetworkLogger from "../../helpers/networkLogger";
-import { responseStatus } from "../../helpers/assertions/index";
-import fixtureFactory from "../../helpers/fixtureFactory";
-import testServerUrl from "../../helpers/constants/testServerUrl";
+import createNetworkLogger from "../helpers/networkLogger";
+import { responseStatus } from "../helpers/assertions/index";
+import fixtureFactory from "../helpers/fixtureFactory";
+import testServerUrl from "../helpers/constants/testServerUrl";
 
-import debugEnabledConfig from "../../helpers/constants/debugEnabledConfig";
-import configureAlloyInstance from "../../helpers/configureAlloyInstance";
+import debugEnabledConfig from "../helpers/constants/debugEnabledConfig";
+import configureAlloyInstance from "../helpers/configureAlloyInstance";
 
 const networkLogger = createNetworkLogger();
 
 fixtureFactory({
-  title: "C2592: Event command sends a request",
+  title: "C2595: Event command passes the org ID on the request",
   url: `${testServerUrl}/alloyTestPage.html`,
   requestHooks: [networkLogger.edgeEndpointLogs]
 });
 
 test.meta({
-  ID: "C2592",
+  ID: "C2595",
   SEVERITY: "P0",
   TEST_RUN: "Regression"
 });
@@ -27,15 +27,15 @@ const triggerAlloyEvent = ClientFunction(() => {
   });
 });
 
-test("Test C2592: Event command sends a request.", async () => {
+test("Test C2595: Event command passes the org ID on the request.", async () => {
   await configureAlloyInstance("alloy", debugEnabledConfig);
   await triggerAlloyEvent();
 
   await responseStatus(networkLogger.edgeEndpointLogs.requests, 200);
-
   await t.expect(networkLogger.edgeEndpointLogs.requests.length).eql(1);
 
   const request = networkLogger.edgeEndpointLogs.requests[0].request.body;
+  const stringifyRequest = JSON.parse(request);
 
-  await t.expect(request).contains('"key":"value"');
+  await t.expect(stringifyRequest.meta.configOverrides.orgId).ok();
 });
