+++ test/functional/specs/C8119.js
@@ -1,26 +1,28 @@
 import { t, Selector, ClientFunction } from "testcafe";
-import testServerUrl from "../../helpers/constants/testServerUrl";
-import fixtureFactory from "../../helpers/fixtureFactory";
-import baseConfig from "../../helpers/constants/baseConfig";
-import addAnchorToBody from "../../helpers/dom/addAnchorToBody";
-import configureAlloyInstance from "../../helpers/configureAlloyInstance";
-import createConsoleLogger from "../../helpers/consoleLogger";
+import testServerUrl from "../helpers/constants/testServerUrl";
+import fixtureFactory from "../helpers/fixtureFactory";
+import baseConfig from "../helpers/constants/baseConfig";
+import addAnchorToBody from "../helpers/dom/addAnchorToBody";
+import configureAlloyInstance from "../helpers/configureAlloyInstance";
+import createConsoleLogger from "../helpers/consoleLogger";
 
 fixtureFactory({
-  title: "C8118: Send event with information about link clicks.",
+  title:
+    "C8119: Does not send event with information about link clicks if disabled.",
   url: `${testServerUrl}/alloyTestPage.html`
 });
 
 test.meta({
-  ID: "C8118",
+  ID: "C8119",
   SEVERITY: "P0",
   TEST_RUN: "Regression"
 });
 
-test("Test C8118: Load page with link. Click link. Verify event.", async () => {
+test("Test C8119: Load page with link. Click link. Verify no event sent.", async () => {
   const getLocation = ClientFunction(() => document.location.href.toString());
   const logger = createConsoleLogger(t, "log");
   const testConfig = {
+    clickCollectionEnabled: false,
     onBeforeEventSend(options) {
       try {
         // eslint-disable-next-line no-console
@@ -41,6 +43,5 @@ test("Test C8118: Load page with link. Click link. Verify event.", async () => {
   await t.click(Selector("#alloy-link-test"));
   await t.expect(getLocation()).contains("blank.html");
   const newMessages = await logger.getNewMessages();
-  const destinationUrl = newMessages[0];
-  await t.expect(destinationUrl).contains("blank.html");
+  await t.expect(newMessages.length).eql(0);
 });
